Initial commit: The Birthday Problem solution: COURSERA JAVA FOR ANDROID COURSE ASSIGNMENT
Introduction:
In probability theory, the birthday problem concerns the probability that, in a set of n randomly chosen people,
some pair of them will have the same birthday. By the pigeonhole principle, the probability reaches 100% when
the number of people reaches 366 (since there are 365 possible birthdays, excluding February 29th). It would
seem that we would need 183 people (half of 365) to reach a 50% probability. However, 99% probability is
reached with just 57 people and 50% probability with just 23 people. These conclusions are based on the
assumption that each day of the year (except February 29) is equally probable for a birthday.

You are to write a program which will simulate the probability that, in a set of n randomly chosen people, some
pair of them will have the same birthday. The code provided to you will get the value of n, the size of the group
of people, and for the number of simulations to be run. For a single simulation, your program should generate a
random birthday for each person in the group of n people and determine whether or not there are two birthdays
that hit the same day. This simulation should be repeated the desired number of times, keeping track of the
number of simulations for which there were two birthdays on the same day. Once all simulations have been
completed, you are to compute the percentage of runs that had two birthdays on the same day and return that
value to the caller.